// front/src/types/api/eat.ts
export type Eat = {
  id: number;
  use_id?: number;
  name: string;
  category: string;
  kcal?: number;
  price?: number;
  eat_type: string;
  eat_timing: string;
  eat_date: string;
  amount?: number;
  unit?: string;
  protein?: number;
  shop?: string;
  discounted?: boolean;
  note?: string;
};